Moved all of the (optional) image renderer code into a separate module.
