+++ dcc-lsystem/src/lib.rs
@@ -255,6 +255,8 @@ pub mod arena;
 pub mod builder;
 #[cfg(feature = "image_renderer")]
 pub mod image;
+#[cfg(feature = "image_renderer")]
+pub mod image_renderer;
 pub mod renderer;
 pub mod system;
 pub mod token;
